pyproject.toml
[project]
name = "livraria"
version = "0.1.0"
description = ""
authors = [
    {name = "GuiDutra21",email = "221021984@aluno.unb.br"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0,<7.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "black (>=25.12.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "flake8 (>=7.3.0,<8.0.0)"
]
